プロトタイプ情報削除機能

+++ app/controllers/prototypes_controller.rb
@@ -48,9 +48,12 @@ class PrototypesController < ApplicationController
   end
 
   def destroy
-    prototype = Prototype.find(params[:id])
-    prototype.destroy
-    redirect_to root_path
+
+    if user_signed_in?
+      prototype = Prototype.find(params[:id])
+      prototype.destroy
+      redirect_to root_path
+    end
   end
 
   private
